% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitprintr.R
\name{knit_print.summary.lme}
\alias{knit_print.summary.lme}
\title{Pretty print contrast table for lme}
\usage{
\method{knit_print}{summary.lme}(object, ...)
}
\arguments{
\item{object}{of class \code{summary.lme}}

\item{...}{\code{caption, moredec, bold_p, intercept_p}}
}
\description{
Custom method for knitr/printr
}
